Move CharacterRoute to utils
